[JENKINS] iOS sdk beta 2014-01-07-16

+++ iOS/release-beta/AVOSCloud.framework/Versions/A/Headers/AVConstants.h
@@ -227,7 +227,7 @@ extern NSInteger const kAVErrorInvalidLinkedSession;
 
 
 typedef void (^PFBooleanResultBlock)(BOOL succeeded, NSError *error);
-typedef void (^PFIntegerResultBlock)(int number, NSError *error);
+typedef void (^PFIntegerResultBlock)(NSInteger number, NSError *error);
 typedef void (^PFArrayResultBlock)(NSArray *objects, NSError *error);
 typedef void (^PFObjectResultBlock)(AVObject *object, NSError *error);
 typedef void (^PFSetResultBlock)(NSSet *channels, NSError *error);
@@ -237,7 +237,7 @@ typedef void (^PFImageResultBlock)(UIImage * image, NSError *error);
 typedef void (^PFDataStreamResultBlock)(NSInputStream *stream, NSError *error);
 typedef void (^PFStringResultBlock)(NSString *string, NSError *error);
 typedef void (^PFIdResultBlock)(id object, NSError *error);
-typedef void (^PFProgressBlock)(int percentDone);
+typedef void (^PFProgressBlock)(NSInteger percentDone);
 typedef void (^PFFileResultBlock)(AVFile * file, NSError *error);
 typedef void (^AVDictionaryResultBlock)(NSDictionary * dict, NSError *error);
